+++ PKHeX.Core/Legality/RNG/ClassicEra/Gen3/MethodH.cs
@@ -606,7 +606,7 @@ public static class MethodH
     {
         if (format == Format && min > 1)
             return level == min; // Met Level matches
-        return LevelRangeExtensions.IsLevelWithinRange((int)level, min, max);
+        return LevelRangeExtensions.IsLevelWithinRange((byte)level, min, max);
     }
 
     public static uint GetRandomLevel<T>(T enc, uint u16LevelRand, LeadRequired lead) where T : ILevelRange => lead switch
