+++ PKHeX.Core/Legality/RNG/ClassicEra/Gen4/MethodJ.cs
@@ -474,7 +474,7 @@ public static class MethodJ
     {
         if (format == Format && min > 1)
             return level == min; // Met Level matches
-        return LevelRangeExtensions.IsLevelWithinRange((int)level, min, max);
+        return LevelRangeExtensions.IsLevelWithinRange((byte)level, min, max);
     }
 
     public static uint GetRandomLevel<T>(T enc, uint seed, LeadRequired lead) where T : IEncounterSlot4
